feat: final changes and challenge concluded
